requirements.txt
fastapi
numpy
onnxruntime
uvicorn[standard]
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
pytest
httpx